requirements.txt
# Speech Emotion Recognition — Windows Compatible Requirements
# Tested on Python 3.10, 3.11, 3.12 on Windows
# Install: pip install -r requirements.txt

tensorflow-cpu==2.15.0
librosa==0.10.1
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
soundfile==0.12.1
audioread==3.0.1
